requirements.txt
# Mortar-C Requirements (Python 3.11+)

# Core
python-dotenv==1.0.1
pydantic==2.9.2

# LLM Backends
anthropic==0.39.0
openai==1.55.3
xai-sdk==0.1.19
grpcio==1.68.1
httpx==0.27.2
requests==2.32.3
anyio==4.7.0

# Blockchain
web3==7.6.0
eth-abi==5.1.0
eth-utils==5.1.0
eth-account==0.13.4
hexbytes==1.2.1
pycryptodome==3.21.0

# Analysis
slither-analyzer==0.10.4
z3-solver==4.13.4.0

# Data Structures
networkx==3.4.2
numpy==2.2.0

# File Handling
filelock==3.16.1
pyyaml==6.0.2

# CLI
rich==13.9.4

# HTTP/Async
sniffio==1.3.1
h11==0.14.0
httpcore==1.0.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1

# Utils
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
annotated-types==0.7.0
pydantic-core==2.27.1
typing-extensions==4.12.2
distro==1.9.0
jiter==0.8.0
tqdm==4.67.1
